Course 5(Advanced System Development)/Workshop 13/TestingAPI/requirements.txt
Flask
bs4
pyzbar
pillow
qrcode[pil]